Moved interface and class into seperate files
