+++ src/main/java/com/vinoth/api/App.java
@@ -1,10 +1,10 @@
 package com.vinoth.api;
 
+import com.vinoth.api.historymanagement.HistoryManagement.InMemoryChatService;
 import com.vinoth.api.model.Calculation;
 import com.vinoth.api.model.Calculation.Operation;
-import com.vinoth.api.model.HistoryManagement;
+import com.vinoth.api.historymanagement.HistoryManagement;
 import spark.ModelAndView;
-import spark.QueryParamsMap;
 import spark.template.handlebars.HandlebarsTemplateEngine;
 
 import java.util.HashMap;
@@ -17,6 +17,7 @@ public class App {
     public static void main(String[] args) {
 
         HandlebarsTemplateEngine templateEngine = new HandlebarsTemplateEngine();
+        InMemoryChatService history = new InMemoryChatService();
 
         get("/", (request, response) -> {
             Float userInput1 = request.queryMap("input1").floatValue();
@@ -46,8 +47,6 @@ public class App {
             String input = request.queryMap("input").value();
             String op = request.queryMap("op").value();
 
-            HistoryManagement.InMemoryChatService history = new HistoryManagement.InMemoryChatService();
-
             if ("Clear History".equals(op)) {
                 history.clearHistory();
                 map.put("history", history.getCompleteHistory());
@@ -56,9 +55,9 @@ public class App {
             else if ("Submit".equals(op)) {
                 if (input != null && !input.isEmpty()) {
                     history.appendMessage(input);
-
-                    map.put("history", history.getCompleteHistory());
                 }
+
+                map.put("history", history.getCompleteHistory());
             }
 
             return templateEngine.render(new ModelAndView(map, "chat.hbs"));
