+++ src/main/java/com/vinoth/api/historymanagement/ChatService.java
@@ -1,0 +1,8 @@
+package com.vinoth.api.historymanagement;
+
+public interface ChatService{
+    void appendMessage(String msg);
+    void appendMessage(String userName, String msg);
+    String getCompleteHistory();
+    void clearHistory();
+}
