+++ src/main/resources/templates/chat.hbs
@@ -5,7 +5,7 @@
         <br><br>
 
         <form>
-            <input type="text" name="input" id="input" width="100">
+            <input type="text" name="input" id="input" width="100" autofocus>
             <br>
             <input type="submit" name="op" value="Submit">
             <input type="submit" name="op" value="Clear History">
